BizWiz2.3/requirements_dynamic.txt
# === DYNAMIC BIZWIZ REQUIREMENTS ===
# Enhanced requirements for real-time data integration

# Core Dash and Plotting
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.17.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# Async and HTTP
aiohttp>=3.8.0
requests>=2.31.0

# Configuration and Serialization
pyyaml>=6.0
pickle-mixin>=1.0.2

# Logging and Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Optional: Advanced Features
# census>=0.8.19          # For real Census API integration
# googlemaps>=4.10.0      # For real Google Places API integration  
# folium>=0.14.0          # Alternative mapping
# geopandas>=0.13.0       # Advanced geographic analysis

# Development and Testing
pytest>=7.4.0
flake8>=6.0.0
black>=23.0.0